Fix HoverContent
