+++ lib/steep.rb
@@ -98,6 +98,7 @@ require "steep/services/content_change"
 require "steep/services/path_assignment"
 require "steep/services/signature_service"
 require "steep/services/type_check_service"
+require "steep/services/hover_content"
 
 require "steep/project"
 require "steep/project/pattern"
@@ -107,7 +108,6 @@ require "steep/project/options"
 require "steep/project/target"
 require "steep/project/dsl"
 require "steep/project/file_loader"
-require "steep/project/hover_content"
 require "steep/project/completion_provider"
 require "steep/project/stats_calculator"
